+++ Samples/Kitchen/DishPrepared.cs
@@ -4,7 +4,7 @@
 
 using Dolittle.SDK.Events;
 
-namespace Basic
+namespace Kitchen 
 {
     [EventType("1844473f-d714-4327-8b7f-5b3c2bdfc26a")]
     public class DishPrepared
@@ -14,7 +14,8 @@ namespace Basic
             Dish = dish;
             Chef = chef;
         }
-        public string Dish { get; set; }
-        public string Chef { get; set; }
+
+        public string Dish { get; }
+        public string Chef { get; }
     }
 }
