+++ Samples/Kitchen/Kitchen.csproj
