+++ Samples/Kitchen/Program.cs
@@ -1,11 +1,11 @@
 ﻿// Copyright (c) Dolittle. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
-// Sample code for the tutorial at https://dolittle.io/tutorial/getting-started/csharp/
+// Sample code for the tutorial at https://dolittle.io/tutorials/getting-started/csharp/
 
 using Dolittle.SDK;
 using Dolittle.SDK.Tenancy;
 
-namespace Basic
+namespace Kitchen
 {
     class Program
     {
